refactor(apes): extract whichBinary into lib/which
