+++ packages/apes/src/commands/agents/destroy.ts
@@ -8,7 +8,7 @@ import { getIdpUrl, loadAuth } from '../../config'
 import { CliError, CliExit } from '../../errors'
 import { apiFetch } from '../../http'
 import { AGENT_NAME_REGEX, buildDestroyTeardownScript, runPhaseGTeardownInProcess } from '../../lib/agent-bootstrap'
-import { whichBinary } from '../../lib/macos-user'
+import { whichBinary } from '../../lib/which'
 import { getHostPlatform, isDarwin } from '../../lib/host-platform'
 import { removeNestAgent } from '../../lib/nest-registry'
 import { readPasswordSilent } from '../../lib/silent-password'
